django_rest_api/requirements.txt
asgiref==3.5.2
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Django==3.2.16
djangorestframework==3.14.0
flake8==5.0.4
idna==3.4
importlib-metadata==4.2.0
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
pycodestyle==2.9.1
pyflakes==2.5.0
pytz==2022.5
requests==2.28.1
sqlparse==0.4.3
tomli==2.0.1
typed-ast==1.5.4
typing_extensions==4.4.0
urllib3==1.26.12
zipp==3.9.0
